backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "FastAPI backend for the Full-Stack Todo Application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "openai-agents>=0.1.0",
    "mcp[cli]>=1.0.0",
    "openai-chatkit>=0.1.0",
    "pydantic-settings>=2.13.1",
    "pyjwt[crypto]>=2.12.1",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.37",
    "uvicorn[standard]>=0.42.0",
]
